backend/requirements.txt
fastapi
uvicorn
pandas
langchain
openai
faiss-cpu
python-dotenv
pydantic